Individual_Key测试完毕

+++ src/Lesson-7/7-1 按键控制LED流水灯模式/lib/individual_key.h
@@ -1,4 +1,5 @@
 #ifndef __LIB_INDIVIDUAL_KEY_H
 #define __LIB_INDIVIDUAL_KEY_H
-int IndividualKey_Scan(void);
+#include "stdint.h"
+int8_t IndividualKey_Scan(void);
 #endif /* __LIB_INDIVIDUAL_KEY_H */
